requirements.txt
openai
langchain
qdrant-client
python-dotenv
pdfplumber
python-docx
pyyaml
tqdm
